pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
bs4 = "^0.0.1"
requests = "^2.27.1"
xlrd = "^2.0.1"
xled = "^0.7.0"
openpyxl = "^3.0.9"
replit = "^3.2.4"
pytz = "^2022.1"
Flask = "^2.1.1"
python-telegram-bot = "^13.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"